requirements.txt
twisted
kaitaistruct
pyOpenSSL
service_identity
click